[#216] merge Repositories into Repository
Also no longer default Elasticsearch index.

Also namespace Elasticsearch test indexes.
